Updated package name and class name for the test
