+++ aws-android-sdk-sagemaker-runtime/src/main/java/com/amazonaws/services/sagemakerruntime/AmazonSageMakerRuntime.java
@@ -13,11 +13,11 @@
  * permissions and limitations under the License.
  */
 
-package com.amazonaws.services.sagemakerrts;
+package com.amazonaws.services.sagemakerruntime;
 
 import com.amazonaws.*;
 import com.amazonaws.regions.*;
-import com.amazonaws.services.sagemakerrts.model.*;
+import com.amazonaws.services.sagemakerruntime.model.*;
 
 /**
  * Interface for accessing Amazon SageMaker Runtime Service
