+++ aws-android-sdk-sagemaker-runtime/src/main/java/com/amazonaws/services/sagemakerruntime/model/package-info.java
@@ -2,5 +2,5 @@
  * Classes modeling the various types represented by AmazonSageMakerRuntime
  */
 
-package com.amazonaws.services.sagemakerrts.model;
+package com.amazonaws.services.sagemakerruntime.model;
