+++ aws-android-sdk-sagemaker-runtime/src/main/java/com/amazonaws/services/sagemakerruntime/model/transform/package-info.java
@@ -2,5 +2,5 @@
  * Marhsalling for the various types represented byAmazonSageMakerRuntime
  */
 
-package com.amazonaws.services.sagemakerrts.model.transform;
+package com.amazonaws.services.sagemakerruntime.model.transform;
