+++ aws-android-sdk-sagemaker-runtime/src/main/java/com/amazonaws/services/sagemakerruntime/package-info.java
@@ -2,5 +2,5 @@
  * <p> The Amazon SageMaker runtime API. </p>
  */
 
-package com.amazonaws.services.sagemakerrts;
+package com.amazonaws.services.sagemakerruntime;
